pyproject.toml
[tool.poetry]
name = "red-crab-inn-bot"
version = "1.1.0"
description = "A discord bot to play Reaper"
authors = ["Agnimandur <shivarasul@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
"discord.py" = "^1.5.1"
replit = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
